requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
Werkzeug==2.3.7
Flet~=0.25.2
requests~=2.32.3